🎨: 장바구니 추가 버튼 클릭 이벤트 핸들러 및 관련 기능 외부 함수로 추출
- CartAddButton 컴포넌트에서 handleAddToCart 이벤트 등록
- handleAddToCart 함수에서 lastSelectedItem 변수를 참조하여 store에 해당 변수 등록 및 업데이트 함수 추가
- CartItem, updateCartItem 함수가 CartAddButton 내에서만 호출되므로 cartAddButton 아래로 위치 변경
