+++ src/basic/main.basic.js
@@ -1,7 +1,5 @@
 import Cart from './components/cart/Cart';
 import CartAddButton from './components/cartAddButton/CartAddButton';
-import CartItem from './components/cartItem/CartItem';
-import { updateCartItemQuantity } from './components/cartItem/updateCartItemQuantity';
 import CartTotal from './components/cartTotal/CartTotal';
 import { renderCalculateCart } from './components/cartTotal/renderCalculateCart';
 import Container from './components/Container';
@@ -13,8 +11,6 @@ import StockStatus from './components/stockStatus/StockStatus';
 import { CONSTANTS } from './constants';
 import { ProductStore } from './store/productStore';
 import { helper } from './utils/helper';
-
-let lastSelectedItem;
 
 function main() {
   const contentWrapper = ContentWrapper();
@@ -34,6 +30,7 @@ function main() {
 
   const productStore = ProductStore.getInstance();
   const products = productStore.getState().products;
+  const lastSelectedItem = productStore.getState().lastSelectedItem;
 
   updateSelectOptions(products);
   renderCalculateCart(products);
@@ -75,48 +72,6 @@ function main() {
 }
 
 main();
-
-// 상품 추가 버튼 클릭 이벤트
-const addToCartButton = document.getElementById('add-to-cart');
-addToCartButton.addEventListener('click', function () {
-  // 선택된 상품 정보 가져오기
-  const productSelector = document.getElementById('product-select');
-  const selectedItem = productSelector.value;
-  const productStore = ProductStore.getInstance();
-  const productToAdd = productStore.findProduct(selectedItem);
-
-  if (productToAdd && productToAdd.quantity > 0) {
-    const productElement = document.getElementById(productToAdd.id);
-
-    if (productElement) {
-      // 이미 장바구니에 있는 상품인 경우 수량 증가
-      const newQuantity =
-        parseInt(
-          productElement.querySelector('span').textContent.split('x ')[1],
-        ) + 1;
-
-      if (newQuantity <= productToAdd.quantity) {
-        // 재고가 충분한 경우 수량 업데이트
-        updateCartItemQuantity(productElement, productToAdd, newQuantity);
-        productStore.updateProductQuantity(productToAdd.id, -1);
-      } else {
-        // 재고가 부족한 경우 알림 표시
-        alert(CONSTANTS.OUT_OF_STOCK_MESSAGE);
-      }
-    } else {
-      // 장바구니에 새로운 상품 추가
-      const newProduct = CartItem(productToAdd);
-      const cart = document.getElementById('cart-items');
-      cart.appendChild(newProduct);
-      productStore.updateProductQuantity(productToAdd.id, -1);
-    }
-
-    // 장바구니 총액 업데이트 및 마지막 선택 상품 기록
-    const products = productStore.getState().products;
-    renderCalculateCart(products);
-    lastSelectedItem = selectedItem;
-  }
-});
 
 // 장바구니 내 상품 조작 이벤트 (수량 변경, 삭제)
 const cart = document.getElementById('cart-items');
